Add cache cloning and catalog update subcommand.
This enables cloning repositories into the local cache based on the
given configuration.

This intentionally relies on execing out to the git client on the users
path (instead of using a library such as go-git) in order to rely on
existing git configuration and credential helpers for authentication and
other settings.

Most of the complexity in this CL comes from the need to mock out exec
calls.

There's some opportunity here to make passing catalog information
cleaner by tying the git funcs to the remote configs themselves, but
unclear if this is something worth doing yet.

Adds repo cloning for #4.