Fix pretty-print header test, base64 encoding for structs.
Pretty printing should use the Content-Type header, not Accept-Encoding
for determining whether the payload contains json or not.

For base64 encoding, the field may be either a primitive value,
(e.g. "a", 1) or a nested JSON struct. We want to handle these
differently to make sure that sjson doesn't string escape nested json,
so test that the decoded type is a valid JSON object and if so, set the
value directly with no escaping.